+++ Libplanet.Tests/Store/StoreTest.cs
@@ -1074,6 +1074,10 @@ namespace Libplanet.Tests.Store
             Assert.Equal(1, store.GetTxNonce(destinationChainId, Fx.Address1));
             Assert.Equal(2, store.GetTxNonce(destinationChainId, Fx.Address2));
             Assert.Equal(3, store.GetTxNonce(destinationChainId, Fx.Address3));
+
+            Assert.Throws<ChainIdNotFoundException>(
+                () => store.ForkTxNonces(Guid.NewGuid(), Guid.NewGuid())
+            );
         }
 
         [SkippableFact]